refactor: remove nodejs module
