+++ lib/Hook.js
@@ -4,7 +4,7 @@
 */
 "use strict";
 
-const util = require("util");
+const util = require("./util");
 
 const deprecateContext = util.deprecate(() => {},
 "Hook.context is deprecated and will be removed");
