+++ lib/HookMap.js
@@ -4,7 +4,7 @@
 */
 "use strict";
 
-const util = require("util");
+const util = require("./util");
 
 const defaultFactory = (key, hook) => hook;
